pdp/requirements.txt
flask==3.0.3
gunicorn==22.0.0
PyJWT==2.8.0
cryptography==43.0.0
